requirements.txt
pandas
numpy
sqlalchemy
psycopg2-binary
scikit-learn
biom-format
h5py
matplotlib
seaborn
scipy